Create LICENSE.md
The code in question is a two sum code, and there are many ideas on the Internet. Because it is simple to implement the function, everyone has no choice but to write code with the same idea, so the codes listed on the Internet are all similar. So I chose the CC0 license because it doesn't require copyright.